modularised notes
